+++ axelor-account/src/main/java/com/axelor/apps/account/service/move/MoveServiceImpl.java
@@ -68,6 +68,7 @@ public class MoveServiceImpl implements MoveService {
   protected MoveExcessPaymentService moveExcessPaymentService;
   protected AccountConfigService accountConfigService;
   protected MoveRepository moveRepository;
+  protected AnalyticMoveLineRepository analyticMoveLineRepository;
 
   protected AppAccountService appAccountService;
 
@@ -84,7 +85,8 @@ public class MoveServiceImpl implements MoveService {
       PaymentService paymentService,
       MoveExcessPaymentService moveExcessPaymentService,
       MoveRepository moveRepository,
-      AccountConfigService accountConfigService) {
+      AccountConfigService accountConfigService,
+      AnalyticMoveLineRepository analyticMoveLineRepository) {
 
     this.moveLineService = moveLineService;
     this.moveCreateService = moveCreateService;
@@ -97,6 +99,7 @@ public class MoveServiceImpl implements MoveService {
     this.moveExcessPaymentService = moveExcessPaymentService;
     this.moveRepository = moveRepository;
     this.accountConfigService = accountConfigService;
+    this.analyticMoveLineRepository = analyticMoveLineRepository;
 
     this.appAccountService = appAccountService;
   }
@@ -591,6 +594,19 @@ public class MoveServiceImpl implements MoveService {
             orgineMoveLine.getCounter(),
             orgineMoveLine.getName(),
             null);
+    reverseMoveLine = this.copyAnalytic(orgineMoveLine, reverseMoveLine);
+    return reverseMoveLine;
+  }
+
+  protected MoveLine copyAnalytic(MoveLine origineMoveLine, MoveLine reverseMoveLine) {
+    reverseMoveLine.setAnalyticDistributionTemplate(
+        origineMoveLine.getAnalyticDistributionTemplate());
+    for (AnalyticMoveLine origineAnalyticMoveLine : origineMoveLine.getAnalyticMoveLineList()) {
+      AnalyticMoveLine copyAnalyticMoveLine =
+          analyticMoveLineRepository.copy(origineAnalyticMoveLine, true);
+      copyAnalyticMoveLine.setMoveLine(reverseMoveLine);
+      reverseMoveLine.addAnalyticMoveLineListItem(copyAnalyticMoveLine);
+    }
     return reverseMoveLine;
   }